Add removing constant slope
